добавляет разделение кода на модули
